k8s: add HPA/v2 for frontend
